Use uint64_t for count

+++ velox/exec/HashAggregation.cpp
@@ -269,7 +269,8 @@ void HashAggregation::resetPartialOutputIfNeed() {
         "flushRowCount", RuntimeCounter(numOutputRows_));
     lockedStats->addRuntimeStat("flushTimes", RuntimeCounter(1));
     lockedStats->addRuntimeStat(
-        "partialAggregationPct", RuntimeCounter(aggregationPct));
+        "partialAggregationPct",
+        RuntimeCounter(static_cast<int64_t>(aggregationPct)));
   }
   groupingSet_->resetTable(/*freeTable=*/false);
   partialFull_ = false;
