+++ velox/experimental/cudf/exec/CudfHashAggregation.cpp
@@ -898,7 +898,8 @@ CudfVectorPtr CudfHashAggregation::releaseAndResetPartialOutput() {
     lockedStats->addRuntimeStat("flushRowCount", RuntimeCounter(numOutputRows));
     lockedStats->addRuntimeStat("flushTimes", RuntimeCounter(1));
     lockedStats->addRuntimeStat(
-        "partialAggregationPct", RuntimeCounter(aggregationPct));
+        "partialAggregationPct",
+        RuntimeCounter(static_cast<int64_t>(aggregationPct)));
   }
 
   numInputRows_ = 0;
